requirements.txt
annotated-types==0.7.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
bcrypt==4.2.1
binaryornot==0.4.4
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
cookiecutter==2.6.0
cryptography==44.0.0
decorator==5.1.1
deprecated==1.2.15
django==5.1.5
django-browser-reload==1.17.0
django-debug-toolbar==5.0.1
django-htmx==1.21.0
django-types==0.20.0
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.1.0
fabric==3.2.2
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
invoke==2.2.0
ipython==8.31.0
jedi==0.19.2
jinja2==3.1.4
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.2
paramiko==3.5.0
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pip==24.3.1
pluggy==1.5.0
prompt-toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.3
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic-core==2.27.2
pygments==2.19.1
pynacl==1.5.0
pyparsing==3.2.1
pyright==1.1.391
pytest==8.3.4
pytest-django==4.9.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-slugify==8.0.4
pyyaml==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
sendgrid==6.11.0
setuptools==75.4.0
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.3
starkbank-ecdsa==2.2.0
stripe==11.4.1
text-unidecode==1.3
tomli==2.0.1
traitlets==5.14.3
types-psycopg2==2.9.21.20241019
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uv==0.5.13
wcwidth==0.2.13
wheel==0.45.1
wrapt==1.17.0
